requirements.txt
transformers==4.40.1
sentence-transformers==2.7.0
accelerate==0.29.3
torch>=2.1.0
faiss-cpu==1.7.4
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.1
tqdm==4.66.2
requests==2.31.0
regex==2024.4.16

